@@ -1,11 +1,10 @@
 name: flutter_shimmer_widget
 description: A package provides an easy way to add shimmer effect in Flutter project 
-version: 1.0.0
+version: 1.1.0
 homepage: https://github.com/daturit/flutter_shimmer_widget
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
